final_version

+++ Data Piplines/plugins/operators/load_dimension.py
@@ -11,12 +11,17 @@ class LoadDimensionOperator(BaseOperator):
                  redshift_conn_id="",
                  table="",
                  insert_statement="",
+                 mode='append_only',
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.table = table
         self.insert_statement = insert_statement
+        if mode == 'append_only':
+            self.append = True
+        else:
+            self.append = False
 
     def execute(self, context):
         """
@@ -26,10 +31,12 @@ class LoadDimensionOperator(BaseOperator):
         # connect redshift
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
 
-        # Clearing data
-        self.log.info(f"Clearing data from destination Redshift table {self.table}")
-        redshift.run("DELETE FROM {}".format(self.table))
+        if not self.append:
+            # Clearing data
+            self.log.info(f"Clearing data from destination Redshift table {self.table}")
+            redshift.run("DELETE FROM {}".format(self.table))
 
         # statement of inserting data
+        self.log.info("Inserting data from staging table!")
         redshift.run(self.insert_statement)
         self.log.info(f"Data Insert Passed on table : {self.table}!!!")
